Split up task runner into task runner and queue.
