+++ memorious/model/__init__.py
@@ -3,3 +3,4 @@ from memorious.model.tag import Tag  # noqa
 from memorious.model.common import Base  # noqa
 from memorious.model.crawler_state import CrawlerState  # noqa
 from memorious.model.crawler_run import CrawlerRun  # noqa
+from memorious.model.queue import Queue  # noqa
